% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hicontrol-package.R
\docType{package}
\name{hicontrol-package}
\alias{hicontrol}
\alias{hicontrol-package}
\title{hicontrol: Control Chart Rules for Influenza HI Titre Data}
\description{
Implements control chart rules adapted for interval-censored influenza
haemagglutination inhibition (HI) titre data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/drserajames/hicontrol}
  \item Report bugs at \url{https://github.com/drserajames/hicontrol/issues}
}

}
\author{
\strong{Maintainer}: Sarah James \email{slj38@cam.ac.uk} (\href{https://orcid.org/0000-0002-6969-1167}{ORCID})

Authors:
\itemize{
  \item Sarah James \email{slj38@cam.ac.uk} (\href{https://orcid.org/0000-0002-6969-1167}{ORCID})
  \item Yi Dong \email{yd332@cantab.ac.uk}
}

}
\keyword{internal}
